services/analysis-worker/requirements.txt
nats-py==2.7.2
structlog==24.1.0
python-dotenv==1.0.1
minio==7.2.7
clickhouse-connect==0.7.8
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
pydantic==2.6.3